# archive/idrm-mvp-docs/40-api-openapi.yaml
openapi: 3.0.3
info:
  title: IDRM API
  description: >-
    Integrated Disaster Response Management — MVP API contract.
    Canonical, incident-centric scheme under /api/v1. URIs are a stable contract
    across the SDLC: the FFP adds endpoints at these same paths, never renaming
    MVP ones. Human-readable companion: 40-api-specification.md.
  version: "1.0.0"
  contact:
    name: IDRM Development Team
  license:
    name: Proprietary (government-endorsed)
servers:
  - url: /api/v1
    description: Version-in-path base; host is environment-specific.
...
security:
  - bearerAuth: []
paths:
...
  /resources:
    get:
      tags: [organizations]
      summary: Provider assets shown on the map (F3)
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: Paged resources.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Resource' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [organizations]
      summary: Declare a provider asset
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Resource' }
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Resource' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }

  # ---------------- Locations / geo (F3, F4) ----------------
  /locations/nearby:
    get:
      tags: [locations]
      summary: PostGIS proximity across incidents and providers
      parameters:
        - { $ref: '#/components/parameters/Latitude' }
        - { $ref: '#/components/parameters/Longitude' }
        - in: query
          name: radius_km
          required: true
          schema: { type: number, minimum: 0.1, maximum: 200 }
        - in: query
          name: layer
          description: Which entities to include.
          schema: { type: string, enum: [all, incidents, providers], default: all }
      responses:
        '200':
          description: Nearby entities, nearest first.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { type: object } }
        '422': { $ref: '#/components/responses/ValidationError' }
  /locations/distance:
    post:
      tags: [locations]
      summary: Distance in km between two points
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [from, to]
              properties:
                from: { $ref: '#/components/schemas/Location' }
                to: { $ref: '#/components/schemas/Location' }
      responses:
        '200':
          description: Distance.
          content:
            application/json:
              schema:
                type: object
                properties: { distance_km: { type: number } }
        '422': { $ref: '#/components/responses/ValidationError' }
  /locations/reverse-geocode:
    get:
      tags: [locations]
      summary: Coordinates to a human-readable address
      parameters:
        - { $ref: '#/components/parameters/Latitude' }
        - { $ref: '#/components/parameters/Longitude' }
      responses:
        '200':
          description: Address.
          content:
            application/json:
              schema:
                type: object
                properties: { address: { type: string } }
        '422': { $ref: '#/components/responses/ValidationError' }
  /locations/map/{layer}:
    parameters:
      - in: path
        name: layer
        required: true
        schema: { type: string, enum: [incidents, providers, resources] }
    get:
      tags: [locations]
      summary: GeoJSON map layer
      responses:
        '200':
          description: A GeoJSON FeatureCollection.
          content:
            application/geo+json:
              schema:
                type: object
                properties:
                  type: { type: string, example: FeatureCollection }
                  features: { type: array, items: { type: object } }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------- Alerts & notifications (F6, F8) ----------------
...
  /notifications:
    get:
      tags: [notifications]
      summary: The user's notifications (F6)
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
        - in: query
          name: unread_only
          description: When true, return only unread notifications.
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Paged notifications.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Notification' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /notifications/unread-count:
    get:
      tags: [notifications]
      summary: Unread badge count
      responses:
        '200':
          description: Count.
          content:
            application/json:
              schema:
                type: object
                properties: { count: { type: integer } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /notifications/{id}/read:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    post:
      tags: [notifications]
      summary: Mark one notification read
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/NotFound' }
  /notifications/read-all:
    post:
      tags: [notifications]
      summary: Mark all read
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /notifications/chat:
    post:
      tags: [notifications]
      summary: FAQ chatbot stub (Task L, PICS-NTF-004)
      description: >-
        Non-intelligent MVP stub — always returns a fixed acknowledgement (no ML).
        Real natural-language help is an FFP intelligence engine (white paper #1).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: { type: string, minLength: 1, maxLength: 2000 }
      responses:
        '200':
          description: Fixed acknowledgement.
          content:
            application/json:
              schema:
                type: object
                properties: { reply: { type: string } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /notifications/{id}:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    delete:
      tags: [notifications]
      summary: Soft-delete a notification
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/NotFound' }
  /notifications/preferences:
    get:
      tags: [notifications]
      summary: Get channel preferences
      responses:
        '200':
          description: Preferences.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationPreferences' }
    patch:
      tags: [notifications]
      summary: Set channel preferences
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NotificationPreferences' }
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationPreferences' }
        '422': { $ref: '#/components/responses/ValidationError' }

  # ---------------- Reports & audit (F8, F9, F10) ----------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
    Page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    Latitude:
      in: query
      name: latitude
      required: true
      schema: { type: number, minimum: -90, maximum: 90 }
    Longitude:
      in: query
      name: longitude
      required: true
      schema: { type: number, minimum: -180, maximum: 180 }
...
  responses:
    Success:
      description: Operation succeeded.
      content:
        application/json:
          schema:
            type: object
            properties: { success: { type: boolean, example: true } }
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but not permitted.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationError:
      description: Semantic validation failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Location:
      type: object
      required: [latitude, longitude]
      properties:
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        total_pages: { type: integer }
    PaginatedEnvelope:
      type: object
      properties:
        data: { type: array, items: {} }
        pagination: { $ref: '#/components/schemas/Pagination' }
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: already_accepted }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  issue: { type: string }
...
    Resource:
      type: object
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        name: { type: string }
        kind: { type: string, example: ambulance }
        quantity: { type: integer }
        location: { $ref: '#/components/schemas/Location' }
...
    Notification:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        message: { type: string }
        read: { type: boolean }
        link: { type: string }
        created_at: { type: string, format: date-time }
    NotificationPreferences:
      type: object
      properties:
        email: { type: boolean }
        sms: { type: boolean }
        in_app: { type: boolean }
...